feat: move EORI call to the proxy
