+++ clients/_test/clientTVA/index.test.ts
@@ -1,4 +1,4 @@
-import { clientTVA } from '#clients/api-vies';
+import { clientTVA } from '#clients/api-proxy/tva';
 import { TVANumber } from '#utils/helpers';
 import { expectClientToMatchSnapshot } from '../expect-client-to-match-snapshot';
